+++ src/streamverifierpk.cpp
@@ -37,7 +37,7 @@ StreamVerifierPK::verify(std::istream &istr,
   // check to see if we've read a final partial chunk
   auto s = istr.gcount();
   if (s != 0) {
-    if (s != plaintext.size())
+    if (static_cast<std::size_t>(s) != plaintext.size())
       plaintext.resize(s);
 
     crypto_sign_update(&state_, plaintext.data(), plaintext.size());
